:sparkles: minions give rewards

+++ src/racks.rs
@@ -6,7 +6,7 @@ use bevy::{
 
 use crate::common::Team;
 
-pub const RACK_GOLD_VALUE: u32 = 10;
+pub const RACK_GOLD_VALUE: f32 = 10.;
 
 #[derive(Component)]
 pub struct Rack {
